Move page metadata to server layouts for client route pages.
Fixes Next.js production build errors when metadata is exported from use client pages.
